performing the subtractions in a wrap-safe way, so that it can work with e.g. incremental encoders that can overflow

+++ examples/DCMotorControllerWithAutoTune/DCMotorControllerWithAutoTune.ino
@@ -70,8 +70,8 @@ const int ONBOARD_LED =	  13;
  *---------------------------------------------------------------------
  */
 
-volatile int32_t positionMeasurement = 0;
-volatile int32_t positionSetpoint = 0; 
+volatile int16_t positionMeasurement = 0;
+volatile int16_t positionSetpoint = 0; 
 uint32_t pidExecutionTimer = 0;
 uint32_t pidExecutionIntervalUs = 1000000 / PID_FREQ_HZ;
 
@@ -259,7 +259,7 @@ void setup() {
  *---------------------------------------------------------------------
  */
 void loop() {
-  int32_t setpoint, measurement;
+  int16_t setpoint, measurement;
   uint32_t now = micros();
 
   if (now - pidExecutionTimer > pidExecutionIntervalUs){
