+++ src/ArduinoPID.cpp
@@ -79,13 +79,14 @@ ConfigError ArduinoPID::getError(){
 	return configError;
 }
 
-int16_t ArduinoPID::compute(int32_t setpoint, int32_t measurement){
+int16_t ArduinoPID::compute(int16_t setpoint, int16_t measurement){
 
     if (configError != NO_ERROR){
         return 0;
     }
 	
-	int32_t err =  int32_t(setpoint) - int32_t(measurement);
+    //wrap safe subtraction to avoid issues with int16_t overflow
+	int32_t err =  (int32_t)(uint16_t(setpoint) - uint16_t(measurement));
 	int64_t output = 0;
 
 	//Calculating Proportional term:
@@ -100,7 +101,7 @@ int16_t ArduinoPID::compute(int32_t setpoint, int32_t measurement){
 	//Calculating Derivative term:
 	if (dGain != 0){
 		if (derivativeFiltering == NO_FILTERING){
-			int32_t deriv = int32_t(measurement) - int32_t(lastMeasurement);
+			int32_t deriv = (int32_t)(uint16_t(measurement) - uint16_t(lastMeasurement));
 			lastMeasurement = measurement;
 			if (deriv > DERIV_MAX){
 				deriv = DERIV_MAX;
